requirements.txt
streamlit==1.30.0
python-dotenv==1.0.0
openai==1.12.0
g4f==0.5.5.5
PyPDF2==3.0.1
python-docx==1.0.1
